requirement.txt
fastapi
starlette
pydantic
uvicorn
langchain
faiss-cpu
transformers
datasets
pymupdf
sentence-transformers
python-dotenv
pdfplumber
google-generativeai
